+++ pyproject.toml
@@ -34,3 +34,7 @@ profile = "black"
 python_version = "3.8"
 ignore_missing_imports = true
 check_untyped_defs = true
+
+[tool.pytest.ini_options]
+pythonpath = ["src"]
+
